+++ src/unifetch/core.nim
@@ -6,7 +6,7 @@ export uri
 
 const uaMozilla* = "Mozilla/5.0 (Windows NT 10.0; rv:109.0) Gecko/20100101 Firefox/115.0"
 
-const unifetchDebugCurl* {.booldefine.} = false
+const showCurlRepr* {.booldefine: "unifetchShowCurlRepr".} = false
 
 type
   UniClientBase* = ref object of RootObj